al fin OUT WARNINGS

+++ TP2/Kernel/VideoDriver.c
@@ -53,7 +53,7 @@ void init_VM_Driver() {
 
 void drawPixel(unsigned int x, unsigned int y, int color)
 {
-    char* screen = (char *) screenData->framebuffer; 
+    char* screen = (char *) (uint64_t)screenData->framebuffer; 
     unsigned where = (x + y*SCREEN_WIDTH) * SCREEN_bPP;
     screen[where] = color & 255;              // BLUE
     screen[where + 1] = (color >> 8) & 255;   // GREEN
